requirements.txt.txt
pandas
mysql-connector-python
matplotlib
seaborn
jupyter
